class

+++ src/214class.kt
@@ -41,7 +41,7 @@ class Rectangle(val height:Int,val width:Int){
 //Kotlin不区分导入的是类还是函数，允许导入任何种类的声明，顶层函数可以直接导入
 import java.util.Random
 
-class Rectangle(val height:Int,val width:Int){
+class Rectangle(val height: Double, val width: Double){
     val isSquare:Boolean
         get() = height == width
 }
@@ -50,3 +50,5 @@ fun createRandomRectangle():Rectangle{
     val  random = Random()
     return Rectangle(random.nextInt(),random.nextInt())
 }
+
+//也可以直接在包名称之后加上.*来导入包中定义的所有声明，这种导入不仅让包中定义的类课件，也会让顶层函数和属性可见。
